ANNMAP-107 Attempting to speed up transcriptToCodingRange
.single.transcript.to.utr.range used to take a single transcript id
and call transcriptDetails on it.

Now, transcriptToUtrRange calls transcriptDetails on all the ids in one go,
and passed each row to .single.transcript.to.utr.range

Also we used to call transcriptToUtrRange for each transcript individually
inside transcriptToCodingRange.  We now call all of them at once, and
get the required rows by indexing on the stable_id of each transcript
in turn